requirements.txt
torch>=1.7
streamlit
Pillow
matplotlib
numpy
